+++ README.md
@@ -25,8 +25,8 @@ The scripts simply wrap the manual commands shown below and are convenient durin
 ### Manual commands
 
 ```bash
-javac -d bin -cp sqlite-jdbc-3.50.1.0.jar src/*.java
-java  -cp bin:sqlite-jdbc-3.50.1.0.jar src.ExpenseTrackerGUI
+javac -d bin -cp sqlite-jdbc-3.50.1.0.jar $(find src -name "*.java")
+java  -cp bin:sqlite-jdbc-3.50.1.0.jar gui.ExpenseTrackerGUI
 ```
 
 ## Database schema
