+++ src/db/DatabaseConnection.java
@@ -1,0 +1,24 @@
+package db;
+
+import java.sql.*;
+
+public class DatabaseConnection {
+    private static final String URL = "jdbc:sqlite:expense_tracker.db";
+
+    public static Connection getConnection() {
+        try {
+            Class.forName("org.sqlite.JDBC");
+            Connection conn = DriverManager.getConnection(URL);
+            try (Statement stmt = conn.createStatement()) {
+                stmt.execute("PRAGMA foreign_keys = ON");
+            }
+            return conn;
+        } catch (ClassNotFoundException e) {
+            System.err.println("JDBC Driver not found: " + e.getMessage());
+            return null;
+        } catch (SQLException e) {
+            System.err.println("Failed to connect to database: " + e.getMessage());
+            return null;
+        }
+    }
+}
