+++ src/model/FoodExpense.java
@@ -1,0 +1,21 @@
+package model;
+
+import java.util.Date;
+
+public class FoodExpense extends Expense {
+    private String restaurant;
+
+    public FoodExpense(double amount, Date date, String description, String restaurant) {
+        super(amount, date, description);
+        this.restaurant = restaurant;
+    }
+
+    @Override
+    public String getType() {
+        return "Food";
+    }
+
+    public String getRestaurant() {
+        return restaurant;
+    }
+}
